Rewrite how Commands and Listeners are specified.
Switch to array declaration (which is less type checked) in order to gain the ability to reload responders. Since it is currently not possible to reload a class definition, I decided to have one responder definition per file in array format.

For commands, one must specify a type=command, identifier, command, parameters, help and answer closure.
For a listener, one must specify a type=listener, identifier, matcher and answer closure.