[#8] Reply to a single request with multiple responses (ResponseBag).
Add a ResponseBag class which will contain 0 to many Response objects in response to a Request.
Update IRC and Shell adapters to handle the ResponseBag.
Update IRC adapter to put its meta in the 'irc' section (instead of directly in the meta).
Responder now respond to a Request and not a string.
Responder now merges the meta into the data given to the Responder callable.
Update Dispatcher so it can reload/reinitialize specific responders.
Update the Reload command to make use of the ability to reload specific responders.