Add some initial support for request preprocessing.
In this particular case, we check if there exist some command alias. If so, we replace the request with the given replacement and execute the request as if it was the request originally sent.

Note: This is currently a dirty hack. It will have to be changed in the future for something better.